refactor(layout): co-locate each node type's sizer with its renderer
